# solution/MainServer/swagger/swaggerDocumentation.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "IumTweb_project",
    "description": "This API is part of the IUM-TWEB project for the 2023-2024 academic year at the Dipartimento di Informatica, Università di Torino. Designed to enable both fans and pundits to access and analyze a wide range of football data, this API supports functionalities across multiple servers (two Express servers and one Java Spring server). It provides interfaces for querying football competitions, clubs, and player statistics, and integrates with MongoDB and Postgres databases to manage dynamic and static datasets respectively. [ IUM-TWEB Group Accornero, Aiello, Bonincontro ]",
    "version": "1.0.0",
    "contact": {
      "name": "IUM-TWEB Group Accornero, Aiello, Bonincontro"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "MainServer - Can handle all API requests."
    },
    {
      "url": "http://localhost:3001",
      "description": "ExpressServer"
    },
    {
      "url": "http://localhost:8081",
      "description": "SpringServer"
    }
  ],
  "paths": {
...
    "/clubs/getByCompetitionAndSeason": {
      "get": {
        "tags": [
          "Clubs"
        ],
        "summary": "Search clubs by competition ID and season",
        "description": "Retrieves a list of clubs participating in a given competition for a specified season.",
        "servers": [
          {
            "url": "http://localhost:3000",
            "description": "MainServer"
          },
          {
            "url": "http://localhost:8081",
            "description": "SpringServer"
          }
        ],
        "parameters": [
          {
            "name": "competition_id",
            "in": "query",
            "required": true,
            "description": "The unique identifier for the competition",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "season",
            "in": "query",
            "required": true,
            "description": "The season for which to retrieve club data",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful - retrieval of club list for the specified competition and season.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Club"
                  }
                }
              }
            }
          },
          "204": {
            "description": "No Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - An error occurred on the server while processing the request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "504": {
            "description": "Gateway Timeout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/clubs/getLastSeason": {
      "get": {
        "tags": [
          "Clubs"
        ],
        "summary": "Retrieve last season for a specified club",
        "servers": [
          {
            "url": "http://localhost:3000",
            "description": "MainServer"
          },
          {
            "url": "http://localhost:8081",
            "description": "SpringServer"
          }
        ],
        "parameters": [
          {
            "name": "club_id",
            "in": "query",
            "required": true,
            "description": "The unique identifier of the club",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful - retrieval of last season for the club.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer"
                }
              }
            }
          },
          "204": {
            "description": "No Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - An error occurred on the server while processing the request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "504": {
            "description": "Gateway Timeout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/clubs/getAllClubsByInitial": {
      "get": {
        "tags": [
          "Clubs"
        ],
        "summary": "Retrieve all clubs grouped by their initial character",
        "servers": [
          {
            "url": "http://localhost:3000",
            "description": "MainServer"
          },
          {
            "url": "http://localhost:8081",
            "description": "SpringServer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful retrieval of clubs grouped by initial character.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Club"
                    }
                  }
                }
              }
            }
          },
          "204": {
            "description": "No Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - An error occurred on the server while processing the request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "504": {
            "description": "Gateway Timeout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/clubs/get": {
      "get": {
        "tags": ["Clubs"],
        "summary": "Get club by ID",
        "description": "Returns information about a club based on club ID.",
        "servers": [
          {
            "url": "http://localhost:3000",
            "description": "MainServer"
          },
          {
            "url": "http://localhost:8081",
            "description": "SpringServer"
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "club_id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID of the club"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful - Returns information about the club with the specified ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Club"
                }
              }
            }
          },
          "204": {
            "description": "No Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - An error occurred on the server while processing the request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "504": {
            "description": "Gateway Timeout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/clubs/getClubsGroupedByInitialAndLikeName/": {
      "get": {
        "tags": ["Clubs"],
        "summary": "Get clubs grouped by initial and name",
        "description": "Returns clubs grouped by initial letter and with names similar to the specified name.",
        "servers": [
          {
            "url": "http://localhost:3000",
            "description": "MainServer"
          },
          {
            "url": "http://localhost:8081",
            "description": "SpringServer"
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Name to search for"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful - Returns clubs grouped by initial letter and with names similar to the specified name.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Club"
                    }
                  }
                }
              }
            }
          },
          "204": {
            "description": "No Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - An error occurred on the server while processing the request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "504": {
            "description": "Gateway Timeout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Club": {
        "type": "object",
        "properties": {
          "clubId": {
            "type": "integer",
            "description": "The unique identifier for the club."
          },
          "clubCode": {
            "type": "string",
            "description": "Unique code of the club."
          },
          "name": {
            "type": "string",
            "description": "Name of the club."
          },
          "domesticCompetitionId": {
            "type": "string",
            "description": "Identifier of the domestic competition the club participates in."
          },
          "totalMarketValue": {
            "type": "number",
            "description": "Total market value of the club."
          },
          "squadSize": {
            "type": "integer",
            "description": "Number of players in the club."
          },
          "averageAge": {
            "type": "number",
            "description": "Average age of the club players."
          },
          "foreignersNumber": {
            "type": "integer",
            "description": "Number of foreign players in the club."
          },
          "foreignersPercentage": {
            "type": "number",
            "description": "Percentage of foreign players in the club."
          },
          "nationalTeamPlayers": {
            "type": "integer",
            "description": "Number of players in the club who are also national team players."
          },
          "stadiumName": {
            "type": "string",
            "description": "Name of the club's stadium."
          },
          "stadiumSeats": {
            "type": "integer",
            "description": "Seating capacity of the stadium."
          },
          "netTransferRecord": {
            "type": "string",
            "description": "Net transfer record of the club."
          },
          "coachName": {
            "type": "string",
            "description": "Name of the club's coach."
          },
          "lastSeason": {
            "type": "string",
            "description": "Performance of the club in the last season."
          },
          "url": {
            "type": "string",
            "description": "URL to more information about the club."
          }
        },
        "description": "Object representing a club"
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": " :( - Error Message"
          },
          "internal_code": {
            "type": "string",
            "description": " :( - Code Error"
          }
        }
      }
    }
  }
}